fixed updated misMascotas

+++ src/main/java/controlador/RegistrarMascotaController.java
@@ -19,7 +19,6 @@ import javax.servlet.http.HttpServletResponse;
 import javax.servlet.http.HttpSession;
 import javax.servlet.http.Part;
 
-
 import modelo.dao.DAOFactory;
 import modelo.entidades.Especie;
 import modelo.entidades.Foto;
@@ -32,7 +31,7 @@ import modelo.entidades.Sexo;
 @MultipartConfig
 public class RegistrarMascotaController extends HttpServlet {
 	private static final long serialVersionUID = 1L;
-	
+
 	private String pathContext = "";
 	private File uploads = new File(pathContext);
 	private String[] extensionesPermitidas = { ".png", ".jpg", ".jpeg" };
@@ -43,20 +42,23 @@ public class RegistrarMascotaController extends HttpServlet {
 		super();
 	}
 
-	protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
+	protected void doGet(HttpServletRequest request, HttpServletResponse response)
+			throws ServletException, IOException {
 		request.getRequestDispatcher("jsp/registrarMascota.jsp").forward(request, response);
 	}
 
-	protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
+	protected void doPost(HttpServletRequest request, HttpServletResponse response)
+			throws ServletException, IOException {
 		String nombre = request.getParameter("nombre");
 		String descripcion = request.getParameter("descripcion");
 		String especie = request.getParameter("especie");
 		String sexo = request.getParameter("sexo");
 		int edad = Integer.parseInt(request.getParameter("edad"));
-		
+
 		HttpSession session = request.getSession();
 		Persona propietario = (Persona) session.getAttribute("usuarioLogeado");
-		Mascota nuevaMascota = new Mascota(nombre, descripcion, Especie.valueOf(especie), Sexo.valueOf(sexo), edad, propietario);
+		Mascota nuevaMascota = new Mascota(nombre, descripcion, Especie.valueOf(especie), Sexo.valueOf(sexo), edad,
+				propietario);
 
 		Part img1 = request.getPart("foto1");
 		Part img2 = request.getPart("foto2");
@@ -66,36 +68,40 @@ public class RegistrarMascotaController extends HttpServlet {
 			fotos.add(new Foto(getUrlFoto(img1), nuevaMascota));
 			fotos.add(new Foto(getUrlFoto(img2), nuevaMascota));
 			fotos.add(new Foto(getUrlFoto(img3), nuevaMascota));
-			
+
 			nuevaMascota.setFotos(fotos);
 			nuevaMascota.setPreferencias(new Preferencias(nuevaMascota, Especie.PERRO, Sexo.MACHO, 1, 15));
 			DAOFactory.getFactory().getMascotaDAO().create(nuevaMascota);
+			List<Mascota> mascotasActualizadas = propietario.getMascotas();
+			mascotasActualizadas.add(nuevaMascota);
+			propietario.setMascotas(mascotasActualizadas);
 		}
-		
+
 		request.getRequestDispatcher("/MisMascotasController").forward(request, response);
 	}
 
 	private String getUrlFoto(Part part) throws ServletException, IOException {
 		String urlResultante = "";
 		this.pathContext = getServletContext().getRealPath("/imgs");
-		
+
 		try {
 			Path path = Paths.get(part.getSubmittedFileName());
 			String fileName = path.getFileName().toString();
 			InputStream input = part.getInputStream();
-			
+
 			if (input != null) {
 				File file = new File(uploads, fileName);
 				Path destino = Paths.get(this.pathContext + "/" + file.toPath());
-				Files.copy(input, destino);
+				if (!Files.exists(destino))
+					Files.copy(input, destino);
 				urlResultante = directorioRaizImg + fileName;
 			}
 		} catch (Exception e) {
 			e.printStackTrace();
 		}
-		
+
 		return urlResultante;
-		
+
 	}
 
 	private boolean extensionValida(String fileName) {
